requirements.txt
opencv-python
opencv-contrib-python

numpy
pandas
matplotlib

iPython
jupyter

torch
torchvision

pillow

tqdm
requests
GitPython